feat(121354): Unifica interfaces IMR
